feat: Added BottomDrawer component
