+++ docs/styleguide.config.js
@@ -64,6 +64,7 @@ module.exports = {
         '../react/Media/Media.jsx',
         '../react/MidEllipsis/index.jsx',
         '../react/Modal/index.jsx',
+        '../react/BottomDrawer/index.jsx',
         '../react/Accordion/index.jsx',
         '../react/Text/index.jsx',
         '../react/Empty/index.jsx',
