+++ react/index.js
@@ -4,6 +4,7 @@ export { Button, ButtonLink } from './Button'
 export { default as ButtonClient } from './PushClientButton'
 export { default as BannerClient } from './PushClientBanner'
 export { default as ButtonAction } from './ButtonAction'
+export { default as BottomDrawer } from './BottomDrawer'
 export { default as BarButton } from './BarButton'
 export { default as I18n, translate, useI18n } from './I18n'
 export { default as Icon, Sprite as IconSprite } from './Icon'
